first push (#20)
* [feat] Add tag and create date to blog list response

* [refactor] Change string author to member author

* [refactor] Change string author to member author

* [feat] Implement to check email service

* [feat] Implement to myInfo

* [feat] Implement member profile image

* [refactor] Combine duplicated codes about images

* [refactor] Delete recipe image service file (#8)

* [feat] Implement to modify password (#9)

* [refactor] Implement interface member service

* [feat] my blog and recipe (#11)

* [feat] Implement my blog list

* [feat] Implement my recipe list

* [feat] Implement to modify blog with images (#12)

* [feat] Implement to modify blog without image

* [feat] Implement to modify blog image

* [feat] Add summary attribute to blog entity (#13)

* [feat] Change member id type (#14)

Long to String

* [feat] Implement to create member nickname if it is null (#15)

* [feat] Implement to modify recipe with image (#16)

After modified, ex-images are still searched in result
However they are removed well in repository and S3

* [feat] blog and recipe delete with id check (#17)

* [feat] Implement global error process (#18)

in member, blog, recipe

* [fix] Fix CORS error of react

* [feat] Add authority attribute in my info (#19)

* [chore] Add cloud data in yml file

+++ src/main/java/com/sweetievegan/blog/service/BlogServiceImp.java
@@ -1,5 +1,7 @@
 package com.sweetievegan.blog.service;
 
+import com.sweetievegan.auth.domain.entity.Member;
+import com.sweetievegan.auth.service.MemberServiceImp;
 import com.sweetievegan.blog.domain.entity.Blog;
 import com.sweetievegan.blog.domain.entity.BlogImage;
 import com.sweetievegan.blog.domain.repository.BlogImageRepository;
@@ -7,8 +9,11 @@ import com.sweetievegan.blog.domain.repository.BlogRepository;
 import com.sweetievegan.blog.dto.request.BlogRegisterRequest;
 import com.sweetievegan.blog.dto.response.BlogDetailResponse;
 import com.sweetievegan.blog.dto.response.BlogListResponse;
-import com.sweetievegan.recipe.domain.entity.RecipeImage;
+import com.sweetievegan.util.exception.GlobalErrorCode;
+import com.sweetievegan.util.exception.GlobalException;
+import com.sweetievegan.util.service.ImageService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -17,12 +22,14 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
+@Slf4j
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class BlogServiceImp implements BlogService {
+	private final MemberServiceImp memberServiceImp;
 	private final BlogRepository blogRepository;
-	private final BlogImageService blogImageService;
+	private final ImageService imageService;
 	private final BlogImageRepository blogImageRepository;
 	@Override
 	public List<BlogListResponse> getAllBlogs() {
@@ -33,7 +40,9 @@ public class BlogServiceImp implements BlogService {
 			BlogListResponse response = BlogListResponse.builder()
 					.id(blog.getId())
 					.title(blog.getTitle())
-					.author(blog.getAuthor())
+					.author(blog.getMember().getNickname())
+					.tag(blog.getTags())
+					.createDate(blog.getCreateDate())
 					.build();
 
 			/* Image files ****************************/
@@ -53,12 +62,16 @@ public class BlogServiceImp implements BlogService {
 	@Override
 	public BlogDetailResponse findBlogByBlogId(Long blogId) {
 		Blog blog = blogRepository.findBlogById(blogId);
+		if(blog == null) {
+			throw new GlobalException(GlobalErrorCode.NOT_FOUND_INFO);
+		}
 		BlogDetailResponse response = BlogDetailResponse.builder()
 				.id(blog.getId())
 				.title(blog.getTitle())
-				.author(blog.getAuthor())
+				.author(blog.getMember().getNickname())
 				.content(blog.getContent())
 				.tags(blog.getTags())
+				.summary(blog.getSummary())
 				.build();
 
 		/* Image files ****************************/
@@ -74,17 +87,20 @@ public class BlogServiceImp implements BlogService {
 	}
 
 	@Override
-	public Long addBlog(BlogRegisterRequest request, List<MultipartFile> file) {
+	public Long addBlog(BlogRegisterRequest request, List<MultipartFile> file, String memberId) {
+		Member member = memberServiceImp.getMemberDetail(memberId);
+
 		Blog blog = Blog.builder()
 				.title(request.getTitle())
-				.author(request.getAuthor())
+				.member(member)
 				.content(request.getContent())
 				.tags(request.getTags())
+				.summary(request.getSummary())
 				.blogImages(new ArrayList<>())
 				.build();
 
 		/* Image files ****************************/
-		List<String> blogImageList = blogImageService.addFile(file, "blog");
+		List<String> blogImageList = imageService.addFile(file, "blog");
 		for(String fn : blogImageList) {
 			blogImageRepository.save(BlogImage.builder()
 					.imageName(fn)
@@ -92,21 +108,59 @@ public class BlogServiceImp implements BlogService {
 					.isDeleted(false)
 					.build());
 		}
-		for (String blogImagePath : blogImageList) {
-			blog.addBlogImage(new BlogImage(blogImagePath));
-		}
 		/* Image files */
 
 		return blogRepository.save(blog).getId();
 	}
 
 	@Override
-	public BlogRegisterRequest updateBlogDetail(Long blogId, BlogRegisterRequest request) {
-		return null;
+	public BlogDetailResponse updateBlogDetail(String memberId, Long blogId, BlogRegisterRequest request, List<MultipartFile> file) {
+		Blog blog = blogRepository.findBlogById(blogId);
+		if(blog == null) {
+			throw new GlobalException(GlobalErrorCode.NOT_FOUND_INFO);
+		}
+		if(!memberId.equals(blog.getMember().getId())) {
+			throw new GlobalException(GlobalErrorCode.NOT_AUTHORIZED_USER);
+		}
+		blog.editBlog(request);
+
+		/* Image files ****************************/
+		/* remove */
+		List<BlogImage> removeBlogImagesList = blog.getBlogImages();
+		for(BlogImage blogImage : removeBlogImagesList) {
+			imageService.removeFile(blogImage.getImageName());
+			blogImageRepository.delete(blogImage);
+		}
+
+		/* add */
+		List<String> blogImageList = imageService.addFile(file, "blog");
+		for(String fn : blogImageList) {
+			blogImageRepository.save(BlogImage.builder()
+					.imageName(fn)
+					.blog(blog)
+					.isDeleted(false)
+					.build());
+		}
+		/* Image files */
+
+		return findBlogByBlogId(blogId);
 	}
 
 	@Override
-	public Long removeBlog(Long blogId) {
+	public Long removeBlog(String memberId, Long blogId) {
+		Blog blog = blogRepository.findBlogById(blogId);
+		if(blog == null) {
+			throw new GlobalException(GlobalErrorCode.NOT_FOUND_INFO);
+		}
+		if(!memberId.equals(blog.getMember().getId())) {
+			throw new GlobalException(GlobalErrorCode.NOT_AUTHORIZED_USER);
+		}
+		List<BlogImage> removeBlogImagesList = blog.getBlogImages();
+		for(BlogImage blogImage : removeBlogImagesList) {
+			imageService.removeFile(blogImage.getImageName());
+			blogImageRepository.delete(blogImage);
+		}
+
 		blogRepository.deleteById(blogId);
 		return blogId;
 	}
